Implement new command system (and renamed old command system to the trigger system, which will later be incorporated into the new command system).
