+++ commands/commands.go
@@ -2,108 +2,133 @@ package commands
 
 import (
 	"fmt"
-	"log"
-	"reflect"
 	"strings"
-
-	"alex-j-butler.com/tf2-booking/util"
 
 	"github.com/bwmarrin/discordgo"
 )
 
-type CommandFunction func(*discordgo.MessageCreate, string, []string)
-
-type CommandHandler struct {
-	function    CommandFunction
-	permissions int
-	respondToDM bool
-}
+type CommandFunction func(*discordgo.MessageCreate, string, []string) bool
 
 type Command struct {
-	Prefix   string
-	Handlers map[string]*CommandHandler
+	// Function to be called when this function is executed.
+	function CommandFunction
+
+	// Specifies the required permissions for the command to run.
+	permissions int
+
+	// Specifies whether this command responds to Direct Messages/Private chat messages.
+	respondToDM bool
+
+	// Subcommands that this command can call.
+	subcommands map[string]*Command
 }
 
-func NewCommand(function CommandFunction) *CommandHandler {
-	return &CommandHandler{
+func NewCommand(function CommandFunction) *Command {
+	return &Command{
 		function:    function,
 		permissions: -1,
-		respondToDM: false,
+		respondToDM: true,
+		subcommands: make(map[string]*Command),
 	}
 }
 
-func (ch *CommandHandler) Permissions(permissions int) *CommandHandler {
-	ch.permissions = permissions
-	return ch
+func (c *Command) Permissions(permissions int) *Command {
+	c.permissions = permissions
+	return c
 }
 
-func (ch *CommandHandler) RespondToDM(respondToDM bool) *CommandHandler {
-	ch.respondToDM = respondToDM
-	return ch
+func (c *Command) RespondToDM(respondToDM bool) *Command {
+	c.respondToDM = respondToDM
+	return c
 }
 
-// New creates a new instance of the Command system
-// with the specified prefix.
-func New(prefix string) *Command {
-	return &Command{
-		Prefix:   prefix,
-		Handlers: make(map[string]*CommandHandler),
+func (c *Command) AddSubcommand(commandName string, command *Command) error {
+	commandName = strings.ToLower(commandName)
+	if _, ok := c.subcommands[commandName]; ok {
+		return fmt.Errorf("%s subcommand already exists", commandName)
 	}
+
+	c.subcommands[commandName] = command
+	return nil
 }
 
-// Add creates a new entry in the command handlers map.
-// First argument accepts a command handler implementing the type `CommandHandler`,
-// Second argument accepts a variable amount of strings specifying the commands to register.
-func (c *Command) Add(handler *CommandHandler, commands ...string) {
-	for _, command := range commands {
-		c.Handlers[command] = handler
+func (c *Command) RemoveSubcommand(commandName string) error {
+	commandName = strings.ToLower(commandName)
+	if _, ok := c.subcommands[commandName]; !ok {
+		return fmt.Errorf("%s subcommand does not exist", commandName)
 	}
+
+	// Delete the subcommand.
+	delete(c.subcommands, commandName)
+	return nil
 }
 
-// Remove deletes entries from the command handlers map.
-func (c *Command) Remove(commands ...string) {
-	for _, command := range commands {
-		delete(c.Handlers, command)
-	}
-}
-
-// Handle the incoming commands and dispatches them to the appropriate
-// command handler, after parsing them.
-func (c *Command) Handle(session *discordgo.Session, m *discordgo.MessageCreate, command string, permissions int) {
-	for str, handler := range c.Handlers {
-		if !strings.HasPrefix(command, c.Prefix) && c.Prefix != "" {
-			continue
+func (c *Command) handleCommand(message *discordgo.MessageCreate, input string, args []string) {
+	// Call the handler function for our command.
+	if c.function != nil {
+		// If the handler function for our command returns true, then we shouldn't handle
+		// anything else here, otherwise we'll continue processing subcommands.
+		if c.function(message, input, args) {
+			return
 		}
+	}
 
-		handlerSplit := strings.Split(str, " ")
-		commandSplit := strings.Split(command[len(c.Prefix):], " ")
+	// If we're here, the handler function was either nil, or returned false informing us we should continue processing
+	// the subcommands.
+	if len(c.subcommands) > 0 && len(args) > 0 {
+		// Check if any subcommands are a match.
+		if command, ok := c.subcommands[strings.ToLower(args[0])]; ok {
+			input = fmt.Sprintf("%s %s", input, args[0])
+			args = args[1:]
 
-		if len(commandSplit) < len(handlerSplit) {
-			continue
+			command.handleCommand(message, input, args)
+			return
 		}
+	}
+}
 
-		if !handler.respondToDM {
-			channel, err := session.State.Channel(m.ChannelID)
-			if err != nil {
-				log.Println("Failed to lookup channel.", err)
-			}
+type CommandSystem struct {
+	// Commands that the command system can handle.
+	commands map[string]*Command
+}
 
-			if channel.IsPrivate {
-				continue
-			}
-		}
+func NewCommandSystem() *CommandSystem {
+	return &CommandSystem{
+		commands: make(map[string]*Command),
+	}
+}
 
-		if reflect.DeepEqual(handlerSplit, commandSplit[:len(handlerSplit)]) {
-			log.Println(fmt.Sprintf("Permissions test: %d & %d = %d", permissions, handler.permissions, permissions&handler.permissions))
+func (cs *CommandSystem) AddCommand(commandName string, command *Command) error {
+	commandName = strings.ToLower(commandName)
+	if _, ok := cs.commands[commandName]; ok {
+		return fmt.Errorf("%s command already exists", commandName)
+	}
 
-			if permissions&handler.permissions != 0 || handler.permissions == -1 {
-				handler.function(m, strings.Join(handlerSplit, " "), commandSplit[len(handlerSplit):])
-			} else {
-				User := &util.PatchUser{m.Author}
-				session.ChannelMessageSend(m.ChannelID, fmt.Sprintf("%s: You don't have permission for that command.", User.GetMention()))
-			}
+	cs.commands[commandName] = command
+	return nil
+}
 
-			break
+func (cs *CommandSystem) RemoveCommand(commandName string) error {
+	commandName = strings.ToLower(commandName)
+	if _, ok := cs.commands[commandName]; !ok {
+		return fmt.Errorf("%s command does not exist", commandName)
+	}
+
+	// Delete the command.
+	delete(cs.commands, commandName)
+	return nil
+}
+
+func (cs *CommandSystem) HandleCommand(message *discordgo.MessageCreate, input string, args []string) {
+	input = args[0]
+
+	if len(cs.commands) > 0 && len(args) > 0 {
+		// Check if any commands are a match.
+		if command, ok := cs.commands[strings.ToLower(input)]; ok {
+			args = args[1:]
+
+			command.handleCommand(message, input, args)
+			return
 		}
 	}
 }
